Major Changes

+++ TheIslandPostManager/ViewModels/DashboardViewModel.cs
@@ -1,10 +1,16 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
+using Microsoft.WindowsAPICodePack.Dialogs;
 using Mysqlx.Crud;
+using System.IO;
+using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Forms;
 using TheIslandPostManager.Dialogs;
+using TheIslandPostManager.Helpers;
 using TheIslandPostManager.Models;
 using TheIslandPostManager.Services;
+using TheIslandPostManager.Views.Pages;
 using TheIslandPostManager.Windows;
 using Wpf.Ui;
 using Wpf.Ui.Controls;
@@ -20,8 +26,9 @@ public partial class DashboardViewModel : ObservableObject
     private readonly IMessageService messageService;
     private readonly IContentDialogService contentDialogService;
     private readonly IMySQLService mySQLService;
+    private readonly INavigationService navigationService;
 
-    public DashboardViewModel(IOrderService orderService,IFileService fileService,IImageService imageService,IMessageService messageService, IContentDialogService contentDialogService,IMySQLService mySQLService)
+    public DashboardViewModel(IOrderService orderService,IFileService fileService,IImageService imageService,IMessageService messageService, IContentDialogService contentDialogService,IMySQLService mySQLService, INavigationService navigationService)
     {
         OrderService = orderService;
         FileService = fileService;
@@ -29,12 +36,20 @@ public partial class DashboardViewModel : ObservableObject
         this.messageService = messageService;
         this.contentDialogService = contentDialogService;
         this.mySQLService = mySQLService;
+        this.navigationService = navigationService;
         OrderService.CreateOrder();
     }
 
     public async Task Open()
     {
-        await ImageService.OpenImageDialogBrowser();
+        try
+        {
+            await ImageService.OpenImageDialogBrowser();
+        }
+        catch (Exception ex)
+        {
+            await messageService.ShowErrorMessage("Load IMages Error", ex.Message, ex.StackTrace, "3c23fd0b-de96-4dcc-a846-3348f26703bb");            
+        }
     }
 
     [RelayCommand]
@@ -45,6 +60,7 @@ public partial class DashboardViewModel : ObservableObject
 
     public async Task CreateOrder()
     {
+        if (OrderService.CurrentOrder is null) return;
 
         var result = await messageService.ShowMessage("Copy Images", "Would you like to copy the images from the current order to this new ?","Cancel",Wpf.Ui.Controls.ControlAppearance.Primary,true,"Yes",true,"No");
 
@@ -90,10 +106,12 @@ public partial class DashboardViewModel : ObservableObject
     }
 
     [RelayCommand]
-    private void DeleteImageClick(ImageObj image)
+    private async Task DeleteImageClick(ImageObj image)
     {
-        ImageService.DeleteImage(image);
-        OrderService.RemoveImageFromOrder(image);
+        if(await ImageService.DeleteImage(image))
+        {
+            OrderService.RemoveImageFromOrder(image);
+        }
     }
 
     [RelayCommand]
@@ -134,19 +152,52 @@ public partial class DashboardViewModel : ObservableObject
 
     internal void OpenCustomerView()
     {
-        CustomerWindow customerWindow = new CustomerWindow(new CustomerWindowViewmodel(OrderService));
-        customerWindow.Show();
+        Screen[] s = Screen.AllScreens;
+
+        if (s.Length > 1)
+        {
+            var display = s.FirstOrDefault(x => x.Primary == false);
+            if (display is not null)
+            {
+                System.Drawing.Rectangle workingArea = display.WorkingArea;
+                CustomerWindow customerWindow = new CustomerWindow(new CustomerWindowViewmodel(OrderService));
+                customerWindow.Left = workingArea.Left;
+                customerWindow.Top = workingArea.Top;
+                customerWindow.Width = workingArea.Width;
+                customerWindow.Height = workingArea.Height;
+                customerWindow.WindowStyle = WindowStyle.None;
+                customerWindow.Topmost = true;
+                customerWindow.Show();
+
+                messageService.ShowSnackBarMessage("Customer View", $"Customer View now visible on screen {display.DeviceName}");
+            }
+        }
+        else
+        {
+            messageService.ShowSnackBarMessage("Customer View", $"Secondary Monitor Not Detected", ControlAppearance.Caution, SymbolRegular.Question48);
+
+        }
+
+
     }
 
     [RelayCommand]
-    private async Task GetCustomerInformation(Order order)
+    private async Task CompleteOrder(Order order)
     {
+        if(order.ApprovedImagesCount <= 0)
+        {
+            messageService.ShowSnackBarMessage("Attention", "No images selected. Please select a photo before attempting to complete this order", ControlAppearance.Caution, SymbolRegular.Stop20);
+            return;
+        }
         OrderService.CurrentOrder = order;
         OrderService.CurrentOrder.IsCompleteingOrder = true;
 
-        var termsOfUseContentDialog = new CompleterOrderDialog(contentDialogService.GetContentPresenter(),OrderService,messageService,mySQLService);
+        //var termsOfUseContentDialog = new CompleterOrderDialog(contentDialogService.GetContentPresenter(),OrderService,messageService,mySQLService);
 
-        _ = await termsOfUseContentDialog.ShowAsync();
+        //await termsOfUseContentDialog.GetEmployees();
+        //_ = await termsOfUseContentDialog.ShowAsync();
+
+        navigationService.Navigate(typeof(CompleteOrderPage));
     }
 
     internal void PreviousPhoto()
@@ -161,14 +212,56 @@ public partial class DashboardViewModel : ObservableObject
 
     internal async Task PendOrder()
     {
-        var termsOfUseContentDialog = new CompletePendingDialog(contentDialogService.GetContentPresenter());
-
-        var result = await termsOfUseContentDialog.ShowAsync();
-
-        if(result == ContentDialogResult.Primary)
+        try
         {
-            var content = ((ComboBoxItem)termsOfUseContentDialog.ViewModel.SelectedItem).Content;
-            await OrderService.PendOrder((string)content);
+            var termsOfUseContentDialog = new CompletePendingDialog(contentDialogService.GetContentPresenter(), mySQLService);
+
+            var result = await termsOfUseContentDialog.ShowAsync();
+
+            if (result == ContentDialogResult.Primary)
+            {
+                var content = ((Employee)termsOfUseContentDialog.ViewModel.SelectedItem);
+                await OrderService.PendOrder((string)content.FullName);
+            }
         }
+        catch (Exception ex)
+        {
+          await  messageService.ShowErrorMessage("Error", ex.Message, ex.StackTrace, "", true);
+        }
+    }
+
+    [RelayCommand]
+    private async Task ReplaceImage(ImageObj imageObj)
+    {
+        var dialog = new CommonOpenFileDialog();
+        dialog.Multiselect = true;
+        dialog.Filters.Add(new CommonFileDialogFilter("JPEG Files", "*.jpg"));
+        dialog.Filters.Add(new CommonFileDialogFilter("PNG Files", "*.png"));
+
+        CommonFileDialogResult result = dialog.ShowDialog();
+
+        if (result == CommonFileDialogResult.Ok)
+        {
+            List<Tuple<string,string>> tuples = new List<Tuple<string,string>>();
+
+            var fileName = Path.GetFileName(dialog.FileName);
+            var settings = App.AppConfig.GetSection("AppSettings") as AppSettings;
+
+            tuples.Add(new Tuple<string, string>(dialog.FileName, Path.Combine(settings.InputDirectory, fileName)));
+
+
+            await Copier.CopyFiles(tuples, (Action<int, string>)((prog, fileName) =>
+            {
+                imageObj.Name = Path.GetFileName(fileName);
+                imageObj.ImageUrl = fileName;
+                imageObj.LoadImage(fileName);
+            }));
+            messageService.ShowSnackBarMessage("Import Finished", "All Images have been imported successfully");
+        }
+    }
+
+    internal void Refresh(string content)
+    {
+        OrderService.CurrentOrder.RefreshImages(content);
     }
 }
